+++ gestiauto-grailsPlugins.iml
@@ -203,6 +203,15 @@
           <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/org.apache.ivy/ivy/jars/ivy-2.3.0.jar!/" />
           <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/org.apache.ant/ant-trax/jars/ant-trax-1.7.1.jar!/" />
           <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/net.java.dev.jna/jna/jars/jna-4.0.0.jar!/" />
+          <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/org.codehaus.jsr166-mirror/jsr166y/jars/jsr166y-1.7.0.jar!/" />
+          <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/org.codehaus.gpars/gpars/jars/gpars-1.2.1.jar!/" />
+          <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/org.hamcrest/hamcrest-core/jars/hamcrest-core-1.3.jar!/" />
+          <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/junit/junit/jars/junit-4.12.jar!/" />
+          <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/org.apache.ant/ant-launcher/jars/ant-launcher-1.9.4.jar!/" />
+          <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/org.apache.ant/ant/jars/ant-1.9.4.jar!/" />
+          <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/org.fusesource.jansi/jansi/jars/jansi-1.11.jar!/" />
+          <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/jline/jline/jars/jline-2.12.jar!/" />
+          <root url="jar://$USER_HOME$/.sdkman/candidates/grails/2.5.5/lib/org.apache.ant/ant-junit/jars/ant-junit-1.9.4.jar!/" />
         </CLASSES>
         <JAVADOC />
         <SOURCES />
