+++ grails-app/controllers/gestion/DashboardController.groovy
@@ -3,7 +3,6 @@ package gestion
 import grails.plugin.springsecurity.annotation.Secured
 
 class DashboardController {
-    
-    @Secured(['ROLE_ADMIN'])
+
     def index() { }
 }
