+++ grails-app/controllers/gestion/RolController.groovy
@@ -3,8 +3,7 @@ package gestion
 import grails.plugin.springsecurity.annotation.Secured
 
 class RolController {
-    
-    @Secured(['ROLE_ADMIN'])
+
     def index() {
      def roles = Rol.findAll()
         println roles.nombreRol
