+++ grails-app/services/gestion/UsuarioService.groovy
@@ -1,0 +1,14 @@
+package gestion
+
+import grails.transaction.Transactional
+
+@Transactional
+class UsuarioService {
+
+    def springSecurityService
+
+    def getTipoUsuarioActual(){
+        def tipoUsuario = springSecurityService.getAuthentication().getAuthorities()
+        return String.valueOf(tipoUsuario)
+    }
+}
