+++ grails-app/views/layouts/_navbar.gsp
@@ -1,11 +1,12 @@
-<link rel="stylesheet" href="/gestion/css/bootstrap.min.css">		
-<link rel="stylesheet" href="/gestion/css/myEstilo.css">
+<link rel="stylesheet" href="/gestion/css/libs/bootstrap.min.css">
+<link rel="stylesheet" href="/gestion/css/libs/myEstilo.css">
+<link rel="stylesheet" href="/gestion/css/libs/jquery-ui.min.css">
 <link rel="stylesheet" href="/gestion/css/libs/fontawesome-all.min.css">
 
 <nav class="navbar navbar-inverse">
     <div class="container-fluid">
         <div class="navbar-header">
-            <a class="navbar-brand" href="${createLink(uri:'/')}">Control Autos</a>
+            <a class="navbar-brand" href="${createLink(uri:'/')}">Control Autos - Bienvenido <sec:username/></a>
         </div>
         <ul class="nav navbar-nav">
