+++ grails-app/views/servicios/citasUsuario.gsp
@@ -1,0 +1,36 @@
+<!DOCTYPE html>
+<html lang="es">
+
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <title>citas terminadas</title>
+</head>
+
+<body>
+<center>
+    <h1>Citas</h1>
+    <g:each in="${servicios}" var="serv">
+        <table>
+            <tr>
+                <th>observaciones</th>
+                <th>modelo</th>
+                <th>marca del auto</th>
+                <th>nombre del usuario</th>
+                <th>estatus</th>
+                <th>acciones</th>
+            </tr>
+            <tr>
+                <td>${serv.observacionesMecanico} </td>
+                <td>${serv.marca.nombreMarca}</td>
+                <td>${serv.automovil.nombreAuto}</td>
+                <td>${serv.usuarios.username}</td>
+                <td>${serv.estatus}</td>
+                <td>
+                    <g:link controller="Servicios" action="delete" id="${serv.id}">eliminar</g:link>
+                </td>
+            </tr>
+        </table>
+    </g:each>
+</center>
+</body>
+</html>
