+++ test/unit/gestion/UsuarioServiceSpec.groovy
@@ -1,0 +1,20 @@
+package gestion
+
+import grails.test.mixin.TestFor
+import spock.lang.Specification
+
+/**
+ * See the API for {@link grails.test.mixin.services.ServiceUnitTestMixin} for usage instructions
+ */
+@TestFor(UsuarioService)
+class UsuarioServiceSpec extends Specification {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+    }
+}
